Codex deduplicated test cases:
Main consolidations:

Interface/property hierarchy cases into interfaces.kt
Getter/setter accessor cases into property_accessors.kt and unique getter/setter cases into unique_fields.kt
Nullability/cast/operator basics into types/nullable.kt, types/casts.kt, and operators/basic.kt
Pure-function locals, branching, and literal returns into local_variables.kt, branching.kt, and literal_returns.kt
Loop coverage into control_flow/loops.kt and user_invariants/loops.kt
Inline/lambda coverage into custom_run_functions.kt and lambdas.kt
