+++ formver.compiler-plugin/testData/diagnostics/verification/classes/interfaces.fir.diag.txt
