+++ formver.compiler-plugin/testData/diagnostics/verification/classes/property_accessors.fir.diag.txt
@@ -1,0 +1,159 @@
+/property_accessors.kt: info: Generated Viper text for testPrimitivePropertyGetter:
+method aProp(this: Ref, anon: Ref) returns (v_ret: Ref)
+
+
+method g_aProp(this: Ref) returns (v_ret: Ref)
+
+
+method nProp(this: Ref) returns (v_ret: Ref)
+
+
+method testPrimitivePropertyGetter(pp: Ref) returns (ret: Ref)
+  requires isSubtype(typeOf(pp), _c_PrimitiveProperty())
+  ensures isSubtype(typeOf(ret), intType())
+{
+  var anon: Ref
+  inhale acc(shared(pp), wildcard)
+  anon := nProp(pp)
+  ret := anon
+  inhale isSubtype(typeOf(ret), intType())
+  goto ret_0
+  label ret_0
+}
+
+/property_accessors.kt: info: Generated Viper text for testPrimitivePropertySetter:
+method aProp(this: Ref, anon: Ref) returns (v_ret: Ref)
+
+
+method g_aProp(this: Ref) returns (v_ret: Ref)
+
+
+method nProp(this: Ref) returns (v_ret: Ref)
+
+
+method testPrimitivePropertySetter(pp: Ref) returns (ret: Ref)
+  requires isSubtype(typeOf(pp), _c_PrimitiveProperty())
+  ensures isSubtype(typeOf(ret), unitType())
+{
+  var anon: Ref
+  inhale acc(shared(pp), wildcard)
+  anon := aProp(pp, intToRef(0))
+  label ret_0
+  inhale isSubtype(typeOf(ret), unitType())
+}
+
+/property_accessors.kt: info: Generated Viper text for testReferencePropertyGetter:
+method aProp(this: Ref, anon_0: Ref) returns (v_ret: Ref)
+
+
+method g_aProp(this: Ref) returns (v_ret: Ref)
+
+
+method g_ppProp(this: Ref) returns (v_ret: Ref)
+
+
+method nProp(this: Ref) returns (v_ret: Ref)
+
+
+method ppProp(this: Ref, anon_0: Ref) returns (v_ret: Ref)
+
+
+method rProp(this: Ref) returns (v_ret: Ref)
+
+
+method testReferencePropertyGetter(rp: Ref) returns (ret: Ref)
+  requires isSubtype(typeOf(rp), _c_ReferenceProperty())
+  ensures isSubtype(typeOf(ret), unitType())
+{
+  var pp: Ref
+  var anon_0: Ref
+  var ppn: Ref
+  var anon: Ref
+  inhale acc(c_ReferenceProperty_shared(rp), wildcard)
+  anon_0 := rProp(rp)
+  pp := anon_0
+  inhale isSubtype(typeOf(pp), _c_PrimitiveProperty())
+  inhale acc(shared(pp), wildcard)
+  anon := nProp(pp)
+  ppn := anon
+  inhale isSubtype(typeOf(ppn), intType())
+  label ret_0
+  inhale isSubtype(typeOf(ret), unitType())
+}
+
+/property_accessors.kt: info: Generated Viper text for testCascadingPropertyGetter:
+method aProp(this: Ref, anon_0: Ref) returns (v_ret: Ref)
+
+
+method g_aProp(this: Ref) returns (v_ret: Ref)
+
+
+method g_ppProp(this: Ref) returns (v_ret: Ref)
+
+
+method nProp(this: Ref) returns (v_ret: Ref)
+
+
+method ppProp(this: Ref, anon_0: Ref) returns (v_ret: Ref)
+
+
+method rProp(this: Ref) returns (v_ret: Ref)
+
+
+method testCascadingPropertyGetter(rp: Ref) returns (ret: Ref)
+  requires isSubtype(typeOf(rp), _c_ReferenceProperty())
+  ensures isSubtype(typeOf(ret), unitType())
+{
+  var ppn: Ref
+  var anon_0: Ref
+  var anon_1: Ref
+  var anon: Ref
+  inhale acc(c_ReferenceProperty_shared(rp), wildcard)
+  anon := rProp(rp)
+  anon_1 := anon
+  inhale isSubtype(typeOf(anon_1), _c_PrimitiveProperty())
+  inhale acc(shared(anon_1), wildcard)
+  anon_0 := nProp(anon_1)
+  ppn := anon_0
+  inhale isSubtype(typeOf(ppn), intType())
+  label ret_0
+  inhale isSubtype(typeOf(ret), unitType())
+}
+
+/property_accessors.kt: info: Generated Viper text for testReferencePropertySetter:
+method aProp(this: Ref, anon_0: Ref) returns (v_ret: Ref)
+
+
+method con() returns (v_ret: Ref)
+  ensures isSubtype(typeOf(v_ret), _c_PrimitiveProperty())
+  ensures acc(shared(v_ret), wildcard)
+  ensures acc(c_PrimitiveProperty_unique(v_ret), write)
+
+
+method g_aProp(this: Ref) returns (v_ret: Ref)
+
+
+method g_ppProp(this: Ref) returns (v_ret: Ref)
+
+
+method nProp(this: Ref) returns (v_ret: Ref)
+
+
+method ppProp(this: Ref, anon_0: Ref) returns (v_ret: Ref)
+
+
+method rProp(this: Ref) returns (v_ret: Ref)
+
+
+method testReferencePropertySetter(rp: Ref) returns (ret: Ref)
+  requires isSubtype(typeOf(rp), _c_ReferenceProperty())
+  ensures isSubtype(typeOf(ret), unitType())
+{
+  var anon_0: Ref
+  var anon: Ref
+  inhale acc(c_ReferenceProperty_shared(rp), wildcard)
+  anon := con()
+  anon_0 := ppProp(rp, anon)
+  label ret_0
+  inhale isSubtype(typeOf(ret), unitType())
+}
