--- formver.compiler-plugin/testData/diagnostics/verification/control_flow/loop.kt
@@ -1,12 +1,0 @@
-
-fun <!VIPER_TEXT!>whileLoop<!>(b: Boolean): Boolean {
-    while (b) {
-        val a = 1
-        val c = 2
-    }
-    return false
-}
-
-fun <!VIPER_TEXT!>whileFunctionCondition<!>() {
-    while (whileLoop(true)) { }
-}
