+++ formver.compiler-plugin/testData/diagnostics/verification/control_flow/loops.fir.diag.txt
