+++ formver.compiler-plugin/testData/diagnostics/verification/control_flow/loops.kt
@@ -1,5 +1,17 @@
-
 import org.jetbrains.kotlin.formver.plugin.NeverConvert
+
+fun <!VIPER_TEXT!>whileLoop<!>(b: Boolean): Boolean {
+    while (b) {
+        val a = 1
+        val c = 2
+    }
+    return false
+}
+
+fun <!VIPER_TEXT!>whileFunctionCondition<!>() {
+    while (whileLoop(true)) {
+    }
+}
 
 @NeverConvert
 fun returnsBoolean(): Boolean {
