+++ formver.compiler-plugin/testData/diagnostics/verification/inlining/custom_run_functions.kt
@@ -8,6 +8,29 @@ import org.jetbrains.kotlin.formver.plugin.AlwaysVerify
 import org.jetbrains.kotlin.formver.plugin.verify
 
 @NeverConvert
+@Suppress("NOTHING_TO_INLINE")
+inline fun double(x: Int): Int {
+    val y = x + x
+    return y
+}
+
+@Suppress("NOTHING_TO_INLINE")
+inline fun <!VIPER_TEXT!>quadruple<!>(x: Int) = double(x) + double(x)
+
+@NeverConvert
+@Suppress("NOTHING_TO_INLINE")
+inline fun branching(b: Boolean): Int {
+    if (b) {
+        return 1
+    } else {
+        return 0
+    }
+}
+
+@Suppress("NOTHING_TO_INLINE")
+inline fun <!VIPER_TEXT!>useBranching<!>(): Int = branching(false) + branching(true)
+
+@NeverConvert
 public inline fun <R> copiedRun(block: () -> R): R = block()
 
 @NeverConvert
