+++ formver.compiler-plugin/testData/diagnostics/verification/inlining/lambdas.fir.diag.txt
@@ -148,3 +148,196 @@ method lambdaPassthrough() returns (v_ret_0: Ref)
   label lbl_ret_0
   inhale isSubtype(typeOf(v_ret_0), unitType())
 }
+
+/lambdas.kt: info: Generated Viper text for captureArg:
+method captureArg(g: Ref) returns (ret_0: Ref)
+  requires isSubtype(typeOf(g), functionType())
+  ensures isSubtype(typeOf(ret_0), intType())
+{
+  var ret_1: Ref
+  var ret: Ref
+  var anon_0: Ref
+  var anon: Ref
+  anon_0 := intToRef(0)
+  inhale isSubtype(typeOf(anon), intType())
+  ret := anon
+  goto ret_2
+  label ret_2
+  ret_1 := ret
+  goto lbl_ret_1
+  label lbl_ret_1
+  ret_0 := ret_1
+  goto lbl_ret_0
+  label lbl_ret_0
+}
+
+/lambdas.kt: info: Generated Viper text for captureVar:
+method captureVar() returns (ret_0: Ref)
+  ensures isSubtype(typeOf(ret_0), intType())
+{
+  var x: Ref
+  var ret_1: Ref
+  var ret: Ref
+  var anon: Ref
+  x := intToRef(1)
+  anon := intToRef(0)
+  ret := plusInts(anon, x)
+  goto ret_2
+  label ret_2
+  ret_1 := ret
+  goto lbl_ret_1
+  label lbl_ret_1
+  ret_0 := ret_1
+  goto lbl_ret_0
+  label lbl_ret_0
+}
+
+/lambdas.kt: info: Generated Viper text for captureAndShadow:
+method captureAndShadow(par_x: Ref) returns (ret_0: Ref)
+  requires isSubtype(typeOf(par_x), intType())
+  ensures isSubtype(typeOf(ret_0), intType())
+{
+  var ret_1: Ref
+  var ret: Ref
+  var anon: Ref
+  var y: Ref
+  var x: Ref
+  anon := intToRef(0)
+  y := par_x
+  x := intToRef(1)
+  ret := plusInts(plusInts(anon, x), y)
+  goto ret_2
+  label ret_2
+  ret_1 := ret
+  goto lbl_ret_1
+  label lbl_ret_1
+  ret_0 := ret_1
+  goto lbl_ret_0
+  label lbl_ret_0
+}
+
+/lambdas.kt: info: Generated Viper text for captureVarClash:
+method captureVarClash(par_x: Ref) returns (ret_0: Ref)
+  requires isSubtype(typeOf(par_x), intType())
+  ensures isSubtype(typeOf(ret_0), intType())
+{
+  var ret_1: Ref
+  var x: Ref
+  var anon_1: Ref
+  var ret: Ref
+  var anon: Ref
+  x := intToRef(1)
+  anon := intToRef(0)
+  ret := timesInts(anon, par_x)
+  goto ret_2
+  label ret_2
+  anon_1 := ret
+  ret_1 := plusInts(anon_1, x)
+  goto lbl_ret_1
+  label lbl_ret_1
+  ret_0 := ret_1
+  goto lbl_ret_0
+  label lbl_ret_0
+}
+
+/lambdas.kt: info: Generated Viper text for captureAndShadowClash:
+method captureAndShadowClash(par_x: Ref) returns (ret_0: Ref)
+  requires isSubtype(typeOf(par_x), intType())
+  ensures isSubtype(typeOf(ret_0), intType())
+{
+  var ret_1: Ref
+  var l1_x: Ref
+  var anon_1: Ref
+  var ret: Ref
+  var anon: Ref
+  var y: Ref
+  var x: Ref
+  l1_x := intToRef(1)
+  anon := intToRef(0)
+  y := par_x
+  x := intToRef(2)
+  ret := plusInts(plusInts(x, y), anon)
+  goto ret_2
+  label ret_2
+  anon_1 := ret
+  ret_1 := plusInts(anon_1, l1_x)
+  goto lbl_ret_1
+  label lbl_ret_1
+  ret_0 := ret_1
+  goto lbl_ret_0
+  label lbl_ret_0
+}
+
+/lambdas.kt: info: Generated Viper text for nestedLambdaShadowing:
+method nestedLambdaShadowing(par_x: Ref) returns (ret_0: Ref)
+  requires isSubtype(typeOf(par_x), intType())
+  ensures isSubtype(typeOf(ret_0), intType())
+{
+  var ret_1: Ref
+  var l1_x: Ref
+  var anon_2: Ref
+  var ret_2: Ref
+  var anon_0: Ref
+  var ret_3: Ref
+  var l3_x: Ref
+  var anon_3: Ref
+  var ret: Ref
+  var anon: Ref
+  var l4_x: Ref
+  var y: Ref
+  var x: Ref
+  l1_x := intToRef(1)
+  anon_0 := intToRef(0)
+  l3_x := intToRef(1)
+  anon := intToRef(0)
+  l4_x := intToRef(3)
+  ret := plusInts(l4_x, anon)
+  goto ret_4
+  label ret_4
+  anon_3 := ret
+  ret_3 := plusInts(anon_3, l3_x)
+  goto lbl_ret_3
+  label lbl_ret_3
+  y := par_x
+  x := intToRef(4)
+  ret_2 := plusInts(plusInts(x, y), anon_0)
+  goto lbl_ret_2
+  label lbl_ret_2
+  anon_2 := ret_2
+  ret_1 := plusInts(anon_2, l1_x)
+  goto lbl_ret_1
+  label lbl_ret_1
+  ret_0 := ret_1
+  goto lbl_ret_0
+  label lbl_ret_0
+}
+
+/lambdas.kt: info: Generated Viper text for callDoubleInvoke:
+method callDoubleInvoke(par_x: Ref) returns (ret_0: Ref)
+  requires isSubtype(typeOf(par_x), intType())
+  ensures isSubtype(typeOf(ret_0), intType())
+{
+  var ret_1: Ref
+  var ret_2: Ref
+  var anon_0: Ref
+  var l2_x: Ref
+  var ret: Ref
+  var anon: Ref
+  var x: Ref
+  anon_0 := intToRef(0)
+  l2_x := anon_0
+  ret_2 := l2_x
+  goto lbl_ret_2
+  label lbl_ret_2
+  anon := intToRef(1)
+  x := anon
+  ret := x
+  goto ret_3
+  label ret_3
+  ret_1 := ret
+  goto lbl_ret_1
+  label lbl_ret_1
+  ret_0 := ret_1
+  goto lbl_ret_0
+  label lbl_ret_0
+}
