--- formver.compiler-plugin/testData/diagnostics/verification/nullability.kt
@@ -1,2 +1,0 @@
-
-fun <!VIPER_TEXT!>return_null<!>(): Int? = null
