+++ formver.compiler-plugin/testData/diagnostics/verification/pure_functions/branching.kt
@@ -1,11 +1,14 @@
 // FULL_JDK
 
+import org.jetbrains.kotlin.formver.plugin.AlwaysVerify
 import org.jetbrains.kotlin.formver.plugin.Pure
 
 @Pure
 fun <!VIPER_TEXT!>noAssignmentInBlocks<!>(a: Boolean): Int {
     var result = 0
-    if (a) { } else { }
+    if (a) {
+    } else {
+    }
     result = result + 1
     return result
 }
@@ -50,9 +53,11 @@ fun <!VIPER_TEXT!>whenExpressionSimple<!>(x: Int): Int {
         1 -> {
             y = 10
         }
+
         2 -> {
             return 20
         }
+
         else -> {
             y = 30
         }
@@ -121,4 +126,46 @@ fun <!VIPER_TEXT!>complexBooleanReturn<!>(x: Int): Boolean {
     } else {
         return true
     }
+}
+
+@AlwaysVerify
+@Pure
+fun <!VIPER_TEXT!>safeDivide<!>(x: Int, y: Int): Int {
+    var res = 0
+    if (y != 0) {
+        res = x / y
+    }
+    return res
+}
+
+@AlwaysVerify
+@Pure
+fun <!VIPER_TEXT!>getStringLength<!>(obj: Any): Int {
+    var len = -1
+    if (obj is String) {
+        len = obj.length
+    }
+    return len
+}
+
+@AlwaysVerify
+@Pure
+fun <!VIPER_TEXT!>safeNestedDivide<!>(x: Int, y: Int, z: Int): Int {
+    var res = 0
+    if (y != 0) {
+        if (z != 0) {
+            res = (x / y) / z
+        }
+    }
+    return res
+}
+
+@AlwaysVerify
+@Pure
+fun <!VIPER_TEXT!>safeInverseDifference<!>(x: Int, y: Int): Int {
+    var res = 0
+    if (x != y) {
+        res = 100 / (x - y)
+    }
+    return res
 }
