+++ formver.compiler-plugin/testData/diagnostics/verification/pure_functions/local_variables.fir.diag.txt
@@ -1,4 +1,71 @@
-/pure_function_with_reassignments.kt: info: Generated Viper text for doubleIncrement:
+/local_variables.kt: info: Generated Viper text for returnNumberVal:
+function returnNumberVal(): Ref
+  ensures isSubtype(typeOf(result), intType())
+{
+  (let l0_x_0 ==
+    (intToRef(42)) in
+    l0_x_0)
+}
+
+/local_variables.kt: info: Generated Viper text for multipleAssignmentsOfDifferentType:
+function multipleAssignmentsOfDifferentType(): Ref
+  ensures isSubtype(typeOf(result), boolType())
+{
+  (let l0_a_0 ==
+    (intToRef(42)) in
+    (let l0_b_0 ==
+      (stringToRef(Seq(72, 101, 108, 108, 111, 32, 83, 110, 97, 75, 116))) in
+      (let l0_c_0 ==
+        (boolToRef(true)) in
+        (let l0_d_0 ==
+          (charToRef(65)) in
+          l0_c_0))))
+}
+
+/local_variables.kt: info: Generated Viper text for multipleAssignmentsWithLiteralReturn:
+function multipleAssignmentsWithLiteralReturn(): Ref
+  ensures isSubtype(typeOf(result), intType())
+{
+  (let l0_a_0 ==
+    (intToRef(42)) in
+    (let l0_b_0 ==
+      (stringToRef(Seq(72, 101, 108, 108, 111, 32, 83, 110, 97, 75, 116))) in
+      (let l0_c_0 ==
+        (boolToRef(true)) in
+        (let l0_d_0 ==
+          (charToRef(65)) in
+          intToRef(42)))))
+}
+
+/local_variables.kt: info: Generated Viper text for laterInitializersCanRelyOnPrevious:
+function laterInitializersCanRelyOnPrevious(): Ref
+  ensures isSubtype(typeOf(result), intType())
+{
+  (let l0_a_0 ==
+    (intToRef(40)) in
+    (let l0_b_0 ==
+      (plusInts(l0_a_0, intToRef(2))) in
+      (let l0_c_0 ==
+        (timesInts(l0_b_0, intToRef(2))) in
+        l0_c_0)))
+}
+
+/local_variables.kt: info: Generated Viper text for initializersCanRelyOnParameters:
+function initializersCanRelyOnParameters(x: Ref, y: Ref): Ref
+  requires isSubtype(typeOf(x), intType())
+  requires isSubtype(typeOf(y), intType())
+  ensures isSubtype(typeOf(result), intType())
+{
+  (let l0_sum_0 ==
+    (plusInts(x, y)) in
+    (let l0_diff_0 ==
+      (minusInts(x, y)) in
+      (let l0_res_0 ==
+        (timesInts(l0_sum_0, l0_diff_0)) in
+        l0_res_0)))
+}
+
+/local_variables.kt: info: Generated Viper text for doubleIncrement:
 function doubleIncrement(): Ref
   ensures isSubtype(typeOf(result), intType())
 {
@@ -11,7 +78,7 @@ function doubleIncrement(): Ref
         x_2)))
 }
 
-/pure_function_with_reassignments.kt: info: Generated Viper text for updateThenReadIntoOther:
+/local_variables.kt: info: Generated Viper text for updateThenReadIntoOther:
 function updateThenReadIntoOther(): Ref
   ensures isSubtype(typeOf(result), intType())
 {
@@ -24,7 +91,7 @@ function updateThenReadIntoOther(): Ref
         y_0)))
 }
 
-/pure_function_with_reassignments.kt: info: Generated Viper text for readOldValueBeforeUpdate:
+/local_variables.kt: info: Generated Viper text for readOldValueBeforeUpdate:
 function readOldValueBeforeUpdate(): Ref
   ensures isSubtype(typeOf(result), intType())
 {
@@ -39,7 +106,7 @@ function readOldValueBeforeUpdate(): Ref
           x_2))))
 }
 
-/pure_function_with_reassignments.kt: info: Generated Viper text for chainThroughTemp:
+/local_variables.kt: info: Generated Viper text for chainThroughTemp:
 function chainThroughTemp(): Ref
   ensures isSubtype(typeOf(result), intType())
 {
@@ -52,7 +119,7 @@ function chainThroughTemp(): Ref
         x_1)))
 }
 
-/pure_function_with_reassignments.kt: info: Generated Viper text for overwriteNotSelfReferential:
+/local_variables.kt: info: Generated Viper text for overwriteNotSelfReferential:
 function overwriteNotSelfReferential(): Ref
   ensures isSubtype(typeOf(result), intType())
 {
@@ -63,7 +130,7 @@ function overwriteNotSelfReferential(): Ref
       x_1))
 }
 
-/pure_function_with_reassignments.kt: info: Generated Viper text for nestedConditionalAssignment:
+/local_variables.kt: info: Generated Viper text for nestedConditionalAssignment:
 function nestedConditionalAssignment(a: Ref, b: Ref): Ref
   requires isSubtype(typeOf(a), boolType())
   requires isSubtype(typeOf(b), boolType())
@@ -88,7 +155,7 @@ function nestedConditionalAssignment(a: Ref, b: Ref): Ref
                 x_0)))))))
 }
 
-/pure_function_with_reassignments.kt: info: Generated Viper text for blockConditionalAssignment:
+/local_variables.kt: info: Generated Viper text for blockConditionalAssignment:
 function blockConditionalAssignment(a: Ref, b: Ref): Ref
   requires isSubtype(typeOf(a), boolType())
   requires isSubtype(typeOf(b), boolType())
@@ -109,7 +176,7 @@ function blockConditionalAssignment(a: Ref, b: Ref): Ref
               x_0))))))
 }
 
-/pure_function_with_reassignments.kt: info: Generated Viper text for whenAssignment:
+/local_variables.kt: info: Generated Viper text for whenAssignment:
 function whenAssignment(a: Ref, b: Ref): Ref
   requires isSubtype(typeOf(a), intType())
   requires isSubtype(typeOf(b), boolType())
