+++ formver.compiler-plugin/testData/diagnostics/verification/types/nullable.fir.diag.txt
@@ -1,3 +1,12 @@
+/nullable.kt: info: Generated Viper text for return_null:
+method return_null() returns (ret: Ref)
+  ensures isSubtype(typeOf(ret), nullable(intType()))
+{
+  ret := nullValue()
+  goto ret_0
+  label ret_0
+}
+
 /nullable.kt: info: Generated Viper text for useNullableTwice:
 method useNullableTwice(x: Ref) returns (v_ret_0: Ref)
   requires isSubtype(typeOf(x), nullable(intType()))
