--- formver.compiler-plugin/testData/diagnostics/verification/types/smartcast.kt
@@ -1,6 +1,0 @@
-
-fun <!VIPER_TEXT!>smartcastReturn<!>(n: Int?): Int =
-    if (n != null) n else 0
-
-fun <!VIPER_TEXT!>isNullOrEmptyWrong<!>(seq: CharSequence?): Boolean =
-    seq == null && seq?.length == 0
