+++ formver.compiler-plugin/testData/diagnostics/verification/user_invariants/loops.fir.diag.txt
@@ -1,4 +1,169 @@
-/while.kt: info: Generated Viper text for test_while:
+/loops.kt: info: Generated Viper text for test:
+field size: Ref
+
+method test(n: Ref) returns (ret_0: Ref)
+  requires isSubtype(typeOf(n), intType())
+  ensures isSubtype(typeOf(ret_0), unitType())
+{
+  var it: Ref
+  var holds: Ref
+  var anon_0: Ref
+  it := intToRef(0)
+  holds := boolToRef(true)
+  label cont_0
+    invariant isSubtype(typeOf(it), intType())
+    invariant isSubtype(typeOf(holds), boolType())
+    invariant isSubtype(typeOf(n), intType())
+    invariant intFromRef(it) <= 10
+    invariant boolFromRef(holds)
+  anon_0 := ltInts(it, intToRef(10))
+  if (boolFromRef(anon_0)) {
+    it := plusInts(it, intToRef(1))
+    goto cont_0
+  }
+  label break_0
+  assert isSubtype(typeOf(it), intType())
+  assert isSubtype(typeOf(holds), boolType())
+  assert isSubtype(typeOf(n), intType())
+  assert intFromRef(it) <= 10
+  assert boolFromRef(holds)
+  assert intFromRef(it) == 10
+  if (intFromRef(it) <= intFromRef(n)) {
+    var anon: Ref
+    label cont
+      invariant isSubtype(typeOf(it), intType())
+      invariant isSubtype(typeOf(holds), boolType())
+      invariant isSubtype(typeOf(n), intType())
+      invariant intFromRef(it) <= intFromRef(n)
+      invariant boolFromRef(holds)
+    anon := ltInts(it, n)
+    if (boolFromRef(anon)) {
+      it := plusInts(it, intToRef(1))
+      goto cont
+    }
+    label break
+    assert isSubtype(typeOf(it), intType())
+    assert isSubtype(typeOf(holds), boolType())
+    assert isSubtype(typeOf(n), intType())
+    assert intFromRef(it) <= intFromRef(n)
+    assert boolFromRef(holds)
+    assert intFromRef(it) == intFromRef(n)
+  }
+  label lbl_ret_0
+  inhale isSubtype(typeOf(ret_0), unitType())
+}
+
+/loops.kt: info: Generated Viper text for loopInsideLoop:
+method loopInsideLoop() returns (ret: Ref)
+  ensures isSubtype(typeOf(ret), unitType())
+{
+  var i: Ref
+  var anon_0: Ref
+  i := intToRef(0)
+  label cont_0
+    invariant isSubtype(typeOf(i), intType())
+    invariant intFromRef(i) <= 10
+  anon_0 := ltInts(i, intToRef(10))
+  if (boolFromRef(anon_0)) {
+    var j: Ref
+    var anon: Ref
+    j := plusInts(i, intToRef(1))
+    label cont
+      invariant isSubtype(typeOf(j), intType())
+      invariant isSubtype(typeOf(i), intType())
+      invariant intFromRef(i) < intFromRef(j)
+      invariant intFromRef(j) <= 10
+    anon := ltInts(j, intToRef(10))
+    if (boolFromRef(anon)) {
+      j := plusInts(j, intToRef(1))
+      goto cont
+    }
+    label break
+    assert isSubtype(typeOf(j), intType())
+    assert isSubtype(typeOf(i), intType())
+    assert intFromRef(i) < intFromRef(j)
+    assert intFromRef(j) <= 10
+    i := plusInts(i, intToRef(1))
+    goto cont_0
+  }
+  label break_0
+  assert isSubtype(typeOf(i), intType())
+  assert intFromRef(i) <= 10
+  label ret_0
+  inhale isSubtype(typeOf(ret), unitType())
+}
+
+/loops.kt: info: Generated Viper text for withBreak:
+field size: Ref
+
+method withBreak() returns (ret_0: Ref)
+  ensures isSubtype(typeOf(ret_0), unitType())
+{
+  var i: Ref
+  var anon: Ref
+  i := intToRef(0)
+  label cont
+    invariant isSubtype(typeOf(i), intType())
+    invariant intFromRef(i) <= 10
+  anon := boolToRef(true)
+  if (boolFromRef(anon)) {
+    if (intFromRef(i) >= 10) {
+      goto break
+    }
+    goto cont
+  }
+  label break
+  assert isSubtype(typeOf(i), intType())
+  assert intFromRef(i) <= 10
+  assert intFromRef(i) == 10
+  label lbl_ret_0
+  inhale isSubtype(typeOf(ret_0), unitType())
+}
+
+/loops.kt: info: Generated Viper text for test_boolean_postcondition:
+field bf_e: Ref
+
+field size: Ref
+
+method con(e: Ref) returns (ret: Ref)
+  requires isSubtype(typeOf(e), intType())
+  ensures isSubtype(typeOf(ret), _c_WithVar())
+  ensures acc(c_WithVar_shared(ret), wildcard)
+  ensures acc(c_WithVar_unique(ret), write)
+
+
+method doSomething(this: Ref) returns (ret: Ref)
+  requires isSubtype(typeOf(this), _c_WithVar())
+  ensures isSubtype(typeOf(ret), boolType())
+
+
+method test_boolean_postcondition() returns (ret_0: Ref)
+  ensures isSubtype(typeOf(ret_0), unitType())
+{
+  var withVar: Ref
+  var boolean: Ref
+  var anon: Ref
+  withVar := con(intToRef(42))
+  boolean := boolToRef(true)
+  label cont
+    invariant acc(c_WithVar_shared(withVar), wildcard)
+    invariant isSubtype(typeOf(withVar), _c_WithVar())
+    invariant isSubtype(typeOf(boolean), boolType())
+  anon := boolean
+  if (boolFromRef(anon)) {
+    boolean := doSomething(withVar)
+    goto cont
+  }
+  label break
+  assert acc(c_WithVar_shared(withVar), wildcard)
+  assert isSubtype(typeOf(withVar), _c_WithVar())
+  assert isSubtype(typeOf(boolean), boolType())
+  assert !boolFromRef(boolean)
+  label lbl_ret_0
+  inhale isSubtype(typeOf(ret_0), unitType())
+}
+
+/loops.kt: info: Generated Viper text for test_while:
 field bf_field: Ref
 
 field size: Ref
@@ -73,7 +238,7 @@ method test_while(param: Ref) returns (v_ret_0: Ref)
   inhale isSubtype(typeOf(v_ret_0), unitType())
 }
 
-/while.kt: info: Generated Viper text for test_while_with_inlining:
+/loops.kt: info: Generated Viper text for test_while_with_inlining:
 field bf_field: Ref
 
 field size: Ref
@@ -164,7 +329,7 @@ method test_while_with_inlining(param: Ref) returns (v_ret_0: Ref)
   inhale isSubtype(typeOf(v_ret_0), unitType())
 }
 
-/while.kt: info: Generated Viper text for test_while_with_smartcast:
+/loops.kt: info: Generated Viper text for test_while_with_smartcast:
 field bf_field: Ref
 
 method test_while_with_smartcast(param: Ref, innerParam: Ref)
