--- formver.compiler-plugin/testData/diagnostics/verification/user_invariants/pure_function_with_literal_return.fir.diag.txt
@@ -1,48 +1,0 @@
-/pure_function_with_literal_return.kt: info: Generated Viper text for annotatedIntLitReturn:
-function annotatedIntLitReturn(par_arg: Ref): Ref
-  requires isSubtype(typeOf(par_arg), intType())
-  requires true
-  requires intFromRef(par_arg) >= 42
-  requires intFromRef(par_arg) <= 42
-  ensures isSubtype(typeOf(result), intType())
-  ensures intFromRef(result) == 42
-  ensures (forall anon: Int ::
-      { anon == intFromRef(result) }
-      anon == intFromRef(result) ==> anon == intFromRef(par_arg))
-{
-  intToRef(42)
-}
-
-/pure_function_with_literal_return.kt: info: Generated Viper text for annotatedBoolLitReturn:
-function annotatedBoolLitReturn(par_arg: Ref): Ref
-  requires isSubtype(typeOf(par_arg), intType())
-  requires intFromRef(par_arg) <= 0
-  requires intFromRef(par_arg) >= 0
-  ensures isSubtype(typeOf(result), boolType())
-  ensures boolFromRef(result)
-  ensures intFromRef(par_arg) == 0
-{
-  boolToRef(true)
-}
-
-/pure_function_with_literal_return.kt: info: Generated Viper text for annotatedCharLitReturn:
-function annotatedCharLitReturn(par_arg: Ref): Ref
-  requires isSubtype(typeOf(par_arg), stringType())
-  requires stringFromRef(par_arg) ==
-    Seq(72, 101, 108, 108, 111, 32, 83, 110, 97, 75, 116)
-  ensures isSubtype(typeOf(result), charType())
-  ensures charFromRef(result) == 65
-{
-  charToRef(65)
-}
-
-/pure_function_with_literal_return.kt: info: Generated Viper text for annotatedStringLitReturn:
-function annotatedStringLitReturn(par_arg: Ref): Ref
-  requires isSubtype(typeOf(par_arg), boolType())
-  requires boolFromRef(par_arg)
-  ensures isSubtype(typeOf(result), stringType())
-  ensures stringFromRef(result) ==
-    Seq(72, 101, 108, 108, 111, 32, 83, 110, 97, 75, 116)
-{
-  stringToRef(Seq(72, 101, 108, 108, 111, 32, 83, 110, 97, 75, 116))
-}
